Fixes error caused by moving file
