+++ src/actions/callAPI.js
@@ -1,6 +1,4 @@
-import { REDUX_WP_API_CALL } from './constants/actions';
-
-export * as types from './constants/actions';
+import { REDUX_WP_API_CALL } from '../constants/actions';
 
 export default (name, request, aditionalParams = {}) => ({
   type: REDUX_WP_API_CALL,
