+++ src/index.js
@@ -2,5 +2,5 @@ import ReduxWPAPI from './ReduxWPAPI';
 
 export default ReduxWPAPI;
 export * as RequestStatus from './constants/requestStatus';
-export callAPI from './callAPI';
+export callAPI from './actions/callAPI';
 export { selectQuery } from './selectors';
